% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny-Server.R
\name{cohortGeneratorServer}
\alias{cohortGeneratorServer}
\title{Cohort Generator Server Module}
\usage{
cohortGeneratorServer(id, qns)
}
\arguments{
\item{id}{Namespace id}

\item{qns}{QueryNamespace object}
}
\value{
Server logic
}
\description{
Internal module server for displaying cohort definitions and counts
}
