% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny-Ui.R
\name{cohortGeneratorUi}
\alias{cohortGeneratorUi}
\title{Cohort Generator UI Module}
\usage{
cohortGeneratorUi(id)
}
\arguments{
\item{id}{Namespace id}
}
\value{
Shiny UI
}
\description{
Internal module for displaying cohort definitions and counts
}
