% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagProcessing.R
\name{cohortTagsToTempTable}
\alias{cohortTagsToTempTable}
\title{Create temporary table for cohort tags in database}
\usage{
cohortTagsToTempTable(cohortTags, connection)
}
\arguments{
\item{cohortTags}{Named list where names are tag names and values are integer vectors of cohort IDs}

\item{connection}{DatabaseConnector connection object}
}
\value{
TRUE if successful, throws error otherwise
}
\description{
Creates a temporary table #cse_cohort_tags with cohort IDs and their tags
}
\keyword{internal}
