% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{create cohorts}
\usage{
createCohorts(executionSettings = NULL, ...)
}
\arguments{
\item{executionSettings}{settings greated with @seealso createExecutionSettings}

\item{...}{settings greated with @seealso createExecutionSettings}
}
\description{
Create cohorts
}
