% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{createResultsDataModel}
\alias{createResultsDataModel}
\title{Create the results data model tables on a database server.}
\usage{
createResultsDataModel(connectionDetails, databaseSchema, tablePrefix = "")
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{databaseSchema}{String schema where database schema lives}

\item{tablePrefix}{(Optional) Use if a table prefix is used before table names (e.g. "cd_")}
}
\description{
Create the results data model tables on a database server.
}
\details{
Only PostgreSQL servers are supported.
}
