% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreStudyDiagnostics.R
\name{executePreStudyDiagnostics}
\alias{executePreStudyDiagnostics}
\title{Generate similarity scores}
\usage{
executePreStudyDiagnostics(executionSettings = NULL, ...)
}
\arguments{
\item{executionSettings}{settings greated with @seealso createExecutionSettings}

\item{...}{settings greated with @seealso createExecutionSettings}
}
\description{
create cosine similarity scores
}
