% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export Results}
\usage{
exportResults(executionSettings = NULL, ...)
}
\arguments{
\item{executionSettings}{settings greated with @seealso createExecutionSettings}

\item{...}{settings greated with @seealso createExecutionSettings}
}
\description{
Create a zip file containing csvs of computed results tables
}
