% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagProcessing.R
\name{flattenCohortTags}
\alias{flattenCohortTags}
\title{Flatten cohort tags to unique cohort IDs}
\usage{
flattenCohortTags(cohortTags)
}
\arguments{
\item{cohortTags}{Named list where names are tag names and values are integer vectors of cohort IDs}
}
\value{
Integer vector of unique cohort IDs
}
\description{
Extracts all unique cohort IDs from a nested list of tags
}
\keyword{internal}
