% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineSimilarity.R
\name{generateSimilarityScores}
\alias{generateSimilarityScores}
\title{Generate similarity scores}
\usage{
generateSimilarityScores(executionSettings = NULL, ...)
}
\arguments{
\item{executionSettings}{settings greated with @seealso createExecutionSettings}

\item{...}{settings greated with @seealso createExecutionSettings}
}
\description{
create cosine similarity scores
}
