% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortDefinitions}
\alias{getCohortDefinitions}
\title{Get Cohort Definitions}
\usage{
getCohortDefinitions(qns)
}
\arguments{
\item{qns}{A QueryNamespace object.}
}
\value{
A data.frame of cohort definitions.
}
\description{
Queries the database for available cohort definitions.
}
\examples{
# qns <- createResultsQueryNamespace(...)
# getCohortDefinitions(qns)
}
