% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortDefinitionsTable}
\alias{getCohortDefinitionsTable}
\title{Get Cohort Definitions Table for a Database}
\usage{
getCohortDefinitionsTable(qns, databaseId, counts = TRUE)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{databaseId}{Character or integer. The database ID(s) to fetch data for.}

\item{counts}{Logical. Whether to include counts (default TRUE).}
}
\value{
A data.frame of cohort definitions and counts.
}
\description{
Returns a table of cohort definitions and counts for a specified database.
}
\examples{
# getCohortDefinitionsTable(qns, databaseId = "CCAE")
}
