% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortGeneratorCounts}
\alias{getCohortGeneratorCounts}
\title{Get Cohort Counts by Cohort Definition ID}
\usage{
getCohortGeneratorCounts(qns, cohortDefinitionId)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{cohortDefinitionId}{Cohort definition ID to get counts for.}
}
\value{
A data.frame of cohort counts by database.
}
\description{
Queries the cg_cohort_count table for counts across databases for a specific cohort
}
\examples{
\dontrun{
# qns <- createResultsQueryNamespace(...)
# getCohortGeneratorCounts(qns, cohortDefinitionId = 123)
}
}
