% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortGeneratorDefinitions}
\alias{getCohortGeneratorDefinitions}
\title{Get Cohort Definitions for Cohort Generator Module}
\usage{
getCohortGeneratorDefinitions(qns)
}
\arguments{
\item{qns}{A QueryNamespace object.}
}
\value{
A data.frame of cohort definitions with id and name.
}
\description{
Queries the cg_cohort_definition table for cohort definitions
}
\examples{
\dontrun{
# qns <- createResultsQueryNamespace(...)
# getCohortGeneratorDefinitions(qns)
}
}
