% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortsByTag}
\alias{getCohortsByTag}
\title{Get Cohorts by tag id}
\usage{
getCohortsByTag(qns, tag)
}
\arguments{
\item{qns}{a query namespace object}

\item{tag}{the tag to filter by}
}
\description{
Get cohorts assigned to a specific tag
}
