% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortsByTagWithCounts}
\alias{getCohortsByTagWithCounts}
\title{Get Cohorts by Tag with Counts}
\usage{
getCohortsByTagWithCounts(qns, tag)
}
\arguments{
\item{qns}{a query namespace object}

\item{tag}{the tag to filter by}
}
\description{
Get cohorts by tag with their counts across all databases (only cohorts with count > 0)
}
