% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortSimilarityScores}
\alias{getCohortSimilarityScores}
\title{Get Cohort Similarity Scores}
\usage{
getCohortSimilarityScores(qns, targetCohortId, weights = NULL)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{targetCohortId}{Integer or character. The target cohort definition ID.}

\item{weights}{optional named vector of numeric weights Demographics, "Medical history" "Presentation" "prior meds" "visit context"
with numeric values to adjust}
}
\value{
A data.frame with similarity scores.
}
\description{
Retrieves similarity scores for all comparators to a target cohort.
}
\examples{
# getCohortSimilarityScores(qns, 101)
}
