% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getDatabaseSimilarityScores}
\alias{getDatabaseSimilarityScores}
\title{Get Database Similarity Scores}
\usage{
getDatabaseSimilarityScores(qns, targetCohortId, databaseIds)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{targetCohortId}{Integer or character. The target cohort definition ID.}

\item{databaseIds}{Character vector of database IDs.}
}
\value{
A data.frame with similarity scores for the specified databases.
}
\description{
Retrieves similarity scores for a target cohort across specific databases.
}
\examples{
# getDatabaseSimilarityScores(qns, 101, c("CCAE", "MDCR"))
}
