% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getDatabaseSources}
\alias{getDatabaseSources}
\title{Get All Database Sources}
\usage{
getDatabaseSources(qns)
}
\arguments{
\item{qns}{A QueryNamespace object.}
}
\value{
A data.frame of all database sources.
}
\description{
Retrieves all database source metadata from the results schema.
}
\examples{
# getDatabaseSources(qns)
}
