% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getDbDataSourcesTable}
\alias{getDbDataSourcesTable}
\title{Get Data Source Table for All Databases}
\usage{
getDbDataSourcesTable(qns, reactableTable = TRUE)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{reactableTable}{Logical. If TRUE (default), returns a Reactable table; otherwise returns a data.frame.}
}
\value{
A Reactable table or data.frame of database sources.
}
\description{
Retrieves metadata about all database sources, optionally as a Reactable table.
}
\examples{
# getDbDataSourcesTable(qns)
}
