% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getPairwiseCovariateData}
\alias{getPairwiseCovariateData}
\title{Get Pairwise Covariate Data}
\usage{
getPairwiseCovariateData(
  qns,
  databaseId,
  cohortDefinitionId1,
  cohortDefinitionId2
)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{databaseId}{Character or integer. The database ID.}

\item{cohortDefinitionId1}{Target cohort definition ID.}

\item{cohortDefinitionId2}{Comparator cohort definition ID.}
}
\value{
A data.frame with covariate means and std. differences.
}
\description{
Retrieves covariate mean and standardized difference for a pair of cohorts in a database.
}
\examples{
# getPairwiseCovariateData(qns, "CCAE", 101, 202)
}
