% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{uploadResults}
\alias{uploadResults}
\title{Upload Results}
\usage{
uploadResults(
  connectionDetails,
  databaseSchema,
  zipFileName = NULL,
  importFilePath = NULL,
  tablePrefix = "",
  ...
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{databaseSchema}{String schema where database schema lives}

\item{zipFileName}{Path to zip file containing results (optional if importFilePath is given)}

\item{importFilePath}{Path to already-extracted results folder (optional if zipFileName is given)}

\item{tablePrefix}{(Optional) Use if a table prefix is used before table names (e.g. "cd_")}

\item{...}{Additional parameters passed to ResultModelManager::uploadResults}
}
\description{
Upload results to a database server from either a zip file or a pre-extracted folder.
}
