% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparatorSelectionExplorer.R
\docType{package}
\name{ComparatorSelectionExplorer-package}
\alias{ComparatorSelectionExplorer}
\alias{ComparatorSelectionExplorer-package}
\title{ComparatorSelectionExplorer: OHDSI Comparator Selection Explorer}
\description{
Utility to select comparators for new user cohort population effect estimation.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/OHDSI/ComparatorSelectionExplorer/issues}
}

}
\author{
\strong{Maintainer}: Jamie Gilbert \email{gilbert@ohdsi.org}

Authors:
\itemize{
  \item Jamie Gilbert \email{gilbert@ohdsi.org}
  \item Justin Bohn
  \item Chris Knoll
  \item Patrick Ryan
  \item David Kern
}

}
\keyword{internal}
