% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparatorSelectionExplorer.R
\docType{package}
\name{ComparatorSelectionExplorer-package}
\alias{ComparatorSelectionExplorer}
\alias{ComparatorSelectionExplorer-package}
\title{ComparatorSelectionExplorer: OHDSI Comparator Selection Explorer}
\description{
Comparator Selection Explorer package - generate cosine similarity scores for pairs of exposure cohorts
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/OHDSI/ComparatorSelectionExplorer/issues}
}

}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Jamie Gilbert
  \item Justin Bohn
  \item Chris Knoll
  \item Patrick Ryan
  \item David Kern
}

}
\keyword{internal}
