% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutionSettings.R
\name{createExecutionSettings}
\alias{createExecutionSettings}
\title{createExecutionSettings}
\usage{
createExecutionSettings(
  connectionDetails,
  connection = NULL,
  databaseName = NULL,
  databaseId = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  incrementalFolder = paste0("incremental_", cdmDatabaseSchema),
  resultsDatabaseSchema,
  cohortDatabaseSchema = resultsDatabaseSchema,
  cohortTable = "cse_cohort",
  tempEmulationSchema = getOption("tempEmulationSchema"),
  cohortDefinitionSet = NULL,
  indicationCohortSubsetDefintions = list(),
  targetCohortIds = NULL,
  cohortCountTable = "cse_cohort_count",
  cohortDefinitionTable = "cse_cohort_definition",
  covariateDefTable = "cse_covariate_ref",
  covariateMeansTable = "cse_covariate_means",
  cosineSimStratifiedTable = "cse_cosine_sim",
  minExposureSize = 1000,
  useBulkCohorts = TRUE,
  logFileLocation = paste0("cse-execution-log-", cdmDatabaseSchema, ".txt"),
  exportDir = tempfile(),
  removeExportDir = TRUE,
  generateCohortDefinitionSet = FALSE,
  exportZipFile = file.path(normalizePath(getwd()), paste0("cse_results_",
    cdmDatabaseSchema, ".zip"))
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector::connectionDetails instance}

\item{connection}{DatabaseConnector::connection (defaults to null). Use for persistent storage of
reference}

\item{databaseName}{Database identifier (string)}

\item{databaseId}{Database identifier integer (optional)}

\item{cdmDatabaseSchema}{String Database schema - where data lives}

\item{vocabularyDatabaseSchema}{standard vocabulary database schema}

\item{incrementalFolder}{folder for storage of incremental results for cohort generation}

\item{resultsDatabaseSchema}{String Database schema - where similarity scores are stored}

\item{cohortDatabaseSchema}{String Database schema - where cohort table is created}

\item{cohortTable}{(optional) cohort table}

\item{tempEmulationSchema}{String DatabaseSchema - temp emulation schema for oracle, bigquery}

\item{cohortDefinitionSet}{CohortGenerator::cohortDefinitionSet - intended to be custom exposures or
indication cohorts}

\item{indicationCohortSubsetDefintions}{List of CohortGenerator::cohortSubsetDefinitions these subsets will be applied
to all cohorts used in this study. See cohortGenerator package documentation for
detailed instructions on creating cohort subsets}

\item{targetCohortIds}{(optional) Integer ids for cohorts to limit cosine similarity calculation to
Must be a valid RxNorm ingredient, ATC class or included in the
cohortDefinitionSet}

\item{cohortCountTable}{(optional) count tabls}

\item{cohortDefinitionTable}{(optional) definitions table}

\item{covariateDefTable}{(optional) where covariate definitions are stored}

\item{covariateMeansTable}{(optional) where covariate means are stored}

\item{cosineSimStratifiedTable}{(optional) where stratified cosine similarity scores are stored}

\item{minExposureSize}{(optional) Minimum number of exposures to be included in cosine similarity
analysis (defaults to 1000).}

\item{useBulkCohorts}{Use the cohort generator bulk rxnorm/atc standard cohort set}

\item{logFileLocation}{(optional) Log file location}

\item{exportDir}{(optional) Folder to store results files in before export (default is tempdir)}

\item{removeExportDir}{(optional) remove the export dir after creating zip files?}

\item{generateCohortDefinitionSet}{Boolean - generate the user specified cohortDefinitionSet}

\item{exportZipFile}{Path to zip file output of project}
}
\value{
executionSettings object
}
\description{
executions settings to be used throughout package for analysis
}
