% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutionSettings.R
\name{createExecutionSettings}
\alias{createExecutionSettings}
\title{createExecutionSettings}
\usage{
createExecutionSettings(
  connectionDetails,
  connection = NULL,
  databaseId,
  incrementalFolder = paste0("incremental_", databaseId),
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema,
  cohortDatabaseSchema = resultsDatabaseSchema,
  cohortTable = "cse_cohort",
  tempEmulationSchema = getOption("tempEmulationSchema"),
  cohortDefinitionSet = NULL,
  exportZipFile
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector::connectionDetails instance}

\item{connection}{DatabaseConnector::connection (or null)}

\item{cdmDatabaseSchema}{String Database schema - where data lives}

\item{resultsDatabaseSchema}{String Database schema - where similarity scores are stored}

\item{cohortDatabaseSchema}{String Database schema - where cohort table is created}

\item{tempEmulationSchema}{String DatabaseSchema - temp emulation schema for oracle, bigquery}

\item{cohortDefinitionSet}{CohortGenerator::cohortDefinitionSet}

\item{exportZipFile}{Path to zip file output of project}

\item{cohortReferenceFilePath}{Path to CSV containing list of exposures}
}
\value{
executionSettings object
}
\description{
executions settings to be used throughout package for analysis
}
