% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.R
\name{execute}
\alias{execute}
\title{Execute Package}
\usage{
execute(executionSettings = NULL, ...)
}
\arguments{
\item{executionSettings}{settings greated with @seealso createExecutionSettings}

\item{...}{settings greated with @seealso createExecutionSettings}
}
\description{
Execute data generation step on a given OMOP CDM - connect, create cohorts and generate data

Data can be transferred with the OhdsiSharing package
}
