% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortDefinitions}
\alias{getCohortDefinitions}
\title{Get Cohort Definitions}
\usage{
getCohortDefinitions(qns, search = NULL, tag = NULL)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{search}{Optional string used to filter cohort names.}

\item{tag}{Optional string used to restrict cohorts to a specific tag.}
}
\value{
A data.frame of cohort definitions.
}
\description{
Queries the database for available cohort definitions.
}
\examples{
# qns <- createResultsQueryNamespace(...)
# getCohortDefinitions(qns)
}
