% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getCohortTags}
\alias{getCohortTags}
\title{Get Cohort Tags}
\usage{
getCohortTags(qns, cohortIds = "")
}
\arguments{
\item{qns}{a query namespace object}

\item{cohortIds}{optional cohort ids}
}
\description{
Get any tags assigned to cohorts
}
