% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsQueries.R
\name{getDbCosineSimilarityTable}
\alias{getDbCosineSimilarityTable}
\title{Get Cosine Similarity Table for a Database}
\usage{
getDbCosineSimilarityTable(
  qns,
  targetCohortId,
  comparatorCohortId,
  databaseId,
  weights = NULL,
  returnReactable = FALSE,
  fmtSim = "\%.3f"
)
}
\arguments{
\item{qns}{A QueryNamespace object.}

\item{targetCohortId}{Target cohort definition ID.}

\item{comparatorCohortId}{Comparator cohort definition ID.}

\item{databaseId}{Character or integer. The database ID.}

\item{weights}{a set of numeric weights > 0 to scale each covariate domain in the cosine similarity calculation (optional)}

\item{returnReactable}{Logical. If TRUE, returns a Reactable table; otherwise returns a data.frame.}

\item{fmtSim}{number of signficant figures to format float output}
}
\value{
A Reactable table or data.frame of cosine similarity values.
}
\description{
Retrieves cosine similarity values for a target and comparator cohort in a specific database.
}
\examples{
# getDbCosineSimilarityTable(qns, 101, 202, "CCAE", TRUE)
}
