% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{uploadResults}
\alias{uploadResults}
\title{Upload Results}
\usage{
uploadResults(
  connectionDetails,
  databaseSchema,
  zipFileName,
  tablePrefix = "",
  importFilpath = tempfile(),
  ...
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{databaseSchema}{String schema where database schema lives}

\item{zipFileName}{Path to zipFile containing results}

\item{tablePrefix}{(Optional) Use if a table prefix is used before table names (e.g. "cd_")}

\item{importFilpath}{file path to export zipped results to before upload (optional - default is temporary)}

\item{...}{Elipsis  - see ResultModelManager::uploadResults}
}
\description{
Upload results to database server
}
